refacto on FMMS
