+++ deepod/__init__.py
@@ -1,4 +1,4 @@
 from deepod.version import __version__
-from . import core, models
+from . import core, models, model_selection
 
-__all__ = ['__version__', 'core', 'models']
+__all__ = ['__version__', 'core', 'models', 'model_selection']
