+++ deepod/model_selection/__init__.py
@@ -1,0 +1,3 @@
+from fmms import FMMS
+
+__all__ = ['FMMS']
